# Speed up extrema 3-50x