@giordano 2025-04-29
Probably this and https://github.com/JuliaLang/julia/pull/58267 would make good benchmarks in https://github.com/JuliaCI/BaseBenchmarks.jl to better control upgrades of LLVM?