@mbauman 2025-04-30
Yeah, interesting.  Looks like some platforms don't have a consistent argument ordering of NaNs.  I'm not sure if that's
* something that we actually need extrema to satisfy; maybe we can relax this test?
* something that we actually guarantee that `min` and `max` must satisfy; maybe this is a separate bug?

Here's the MWE on a skylake-avx512 machine:

```
julia> f((min1, max1), (min2, max2)) = (min(min1, min2), max(max1, max2));

julia> x = -NaN, -NaN;

julia> y = NaN, NaN;

julia> signbit.(f(x, y))
(false, true)
```

The test asserts that the signs returned by `f` here are the same — that it either always returns the first arguments to min/max or the second arguments.  This doesn't happen on v1.11, but does on beta2 and master.