@oscardssmith 2025-04-30
In general LLVM (and therefore we) make no guarentees about what NaN you will get.