@mbauman 2025-04-30
OK, great, then we can just relax that test.  I suspect it was written trying to allow for any ordering, but it missed the heterogeneous case.