@oscardssmith 2025-05-02
no. It's a performance improvement, not a bugfix.