@mbauman 2025-05-02
This also isn't uniformly a win on Julias v1.11 and prior — probably because it's standing atop the same compiler change(s) that made #58267 fast ([as Oscar notes](https://github.com/JuliaLang/julia/pull/58267#issuecomment-2847262028), that's likely largely #56371).